split graphical/window & graphical/gui, update doc #17
